requirements.txt
flask>=3.0
requests>=2.31
icalendar>=6.0
python-dateutil>=2.8
